api: change integer schema parameters to i64
The type for the minimum / maximum / default values for integer
schemas has changed to i64 (from isize). Fix all call sites that are
converting to isize to convert to i64 instead.

+++ proxmox-backup-client/src/main.rs
@@ -726,7 +726,7 @@ fn spawn_catalog_upload(
                 type: Integer,
                 description: "Max number of entries to hold in memory.",
                 optional: true,
-                default: pbs_client::pxar::ENCODER_MAX_ENTRIES as isize,
+                default: pbs_client::pxar::ENCODER_MAX_ENTRIES as i64,
             },
             "dry-run": {
                 type: Boolean,
